+++ nuclia/lib/nua_responses.py
@@ -274,9 +274,9 @@ class PublicPushResponse(BaseModel):
 
 
 class ProcessingStatusInfo(BaseModel):
-    last_delivered_seqid: Optional[
-        int
-    ] = None  # When none, means we already don't have information about this queue
+    last_delivered_seqid: Optional[int] = (
+        None  # When none, means we already don't have information about this queue
+    )
 
 
 class ProcessingStatus(BaseModel):
@@ -453,3 +453,30 @@ class StoredLearningConfiguration(BaseModel):
     summary: str
     summary_model: str
     summary_prompt: Optional[SummaryPrompt] = None
+
+
+class SentenceSearch(BaseModel):
+    data: List[float] = []
+    time: float
+
+
+class Ner(BaseModel):
+    text: str
+    ner: str
+    start: int
+    end: int
+
+
+class TokenSearch(BaseModel):
+    tokens: List[Ner] = []
+    time: float
+
+
+class QueryInfo(BaseModel):
+    language: str
+    stop_words: List[str]
+    semantic_threshold: float
+    visual_llm: bool
+    max_context: int
+    entities: Optional[TokenSearch]
+    sentence: Optional[SentenceSearch]